requirements.txt
# Core
feedparser>=6.0.10
requests>=2.31.0
python-dotenv>=1.0.0

# Data & storage
sqlite-utils>=3.35

# LLM integrations
openai>=1.12.0
anthropic>=0.20.0

# Market data
yfinance>=0.2.36

# Web server
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9
markdown>=3.5.0

# NLP / dedup
scikit-learn>=1.4.0
